Writing our first test
there’s a little bug in the polls application for us to fix right away: the Question.was_published_recently() method returns True if the Question was published within the last day (which is correct) but also if the Question’s pub_date field is in the future (which certainly isn’t).

Create a test.py that test above bug and other cases.
command:
    python manage.py test polls